Refactor packages and classes
