+++ src/main/java/school/faang/sprint_1/task_43607/Main.java
@@ -1,0 +1,11 @@
+package school.faang.sprint_1.task_43607;
+
+import java.util.Arrays;
+
+public class Main {
+    public static void main(String[] args) {
+        int[] nums = new int[]{1, 2, 3, 4, 5};
+        Example.reverse(nums);
+        System.out.println(Arrays.toString(nums));
+    }
+}
