+++ src/main/java/school/faang/sprint_1/task_43658/Main.java
@@ -1,4 +1,4 @@
-package school.faang.task_43658;
+package school.faang.sprint_1.task_43658;
 
 public class Main {
     public static void main(String[] args) {
